requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-dotenv==0.19.0
httpx==0.24.1
